+++ priv/templates/components/breadcrumb.exs
@@ -3,20 +3,20 @@
     name: "breadcrumb",
     args: [
       color: [
-        "white",
+        "natural",
         "primary",
         "secondary",
-        "dark",
         "success",
         "warning",
         "danger",
         "info",
-        "light",
+        "silver",
         "misc",
         "dawn"
       ],
       size: ["extra_small", "small", "medium", "large", "extra_large"],
       only: ["breadcrumb"],
+      helpers: [],
       module: ""
     ],
     optional: [],