+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add dark mode of Alert component [#99](https://github.com/mishka-group/mishka_chelekom/pull/99)
 - Add dark mode of Spinner component [#100](https://github.com/mishka-group/mishka_chelekom/pull/100)
 - Add dark mode of Banner component [#101](https://github.com/mishka-group/mishka_chelekom/pull/101)
+- Add dark mode of Breadcrumb component [#102](https://github.com/mishka-group/mishka_chelekom/pull/102)
 
 ### Bugs:
 - Fix un-CSP progress mounted [#72](https://github.com/mishka-group/mishka_chelekom/pull/72)